# Broadcast performance: fix type stability