@mbauman 2018-05-01
Sure, I think this should show improvements in GoGame and Laplacian.

@nanosoldier `runbenchmarks(ALL, vs = ":master")`