@nanosoldier 2018-05-02
[Your benchmark job](https://github.com/JuliaLang/julia/pull/26942#issuecomment-385769511) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/6ba524e0075beecaf85e560eb2c5f269c0999078/d7959e5_vs_23388ae/report.md). cc @ararslan