@mbauman 2018-05-02
Looks like I'll need to add these examples to BaseBenchmarks.